client-cli/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pvc"
version = "1.0.1"
dependencies = [
    "Click",
    "PyYAML",
    "lxml",
    "colorama",
    "requests",
    "requests-toolbelt",
]

[tool.setuptools]
packages = ["pvc.cli", "pvc.lib"]

[project.scripts]
pvc = "pvc.cli.cli:cli"
